+++ app/Http/Controllers/SearchController.php
@@ -4,6 +4,7 @@ namespace KushyApi\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
+use Spatie\QueryBuilder\QueryBuilder;
 use KushyApi\Http\Controllers\Controller;
 use KushyApi\Http\Resources\Shops as ShopsResource;
 use KushyApi\Http\Resources\SearchCollection;
@@ -19,6 +20,35 @@ class SearchController extends Controller
         $this->AddPostCategories = $AddPostCategories;
         $this->CreatePostSlug = $CreatePostSlug;
         $this->UploadPostMedia = $UploadPostMedia;
+    }
+
+    public function index(Request $request)
+    {
+        $config = Config::get('api');
+
+        /**
+         * We use Spatie's Query Builder package to handle
+         * filtering, sorting, and includes
+         */
+
+        $search = QueryBuilder::for(Posts::class)
+            ->with('categories')
+            ->allowedFilters([
+                'post_id', 
+                'user_id', 
+                'rating', 
+                'review'
+            ])
+            ->allowedIncludes([
+                'user', 
+                'post', 
+                'strainMeta'
+            ])
+            ->paginate($config['query']['pagination']);
+        
+        return (new SearchCollection($search))
+            ->response()
+            ->setStatusCode(201);
     }
 
     /**